refactor(suite): split ForgetDevice into a folder with separate modules
Break the monolithic ForgetDevice.tsx into focused files:
- ForgetDevice.tsx — settings entry point
- ForgetDeviceModal.tsx — state resolver and flow picker
- ForgetDeviceFlows.tsx — all flow components
- ConfirmationModal.tsx, UnplugDeviceModal.tsx, etc. — shared modals
- useForgetDevice.ts — hook wrapping forgetDeviceThunk
