+++ packages/suite/src/views/settings/SettingsDevice/ForgetDevice/index.tsx
@@ -1,0 +1,2 @@
+export { ForgetDevice } from './ForgetDevice';
+export { ForgetDeviceModal } from './ForgetDeviceModal';
